Improve getEventsSince() logging
Seems like the previous system state is newer than the current one.

With this change in place, when that happens we include both the previous
state's timestamp and the current state's timestamp in the exception message.
And at least part of the system timestamp is included as part of the log
message.