Remove Androidplot dependency
It's been nothing but trouble.

Of course this makes the whole thing not build any more, so we need to replace
Androidplot with something.

My current thinking is DIY; our needs with the vertical reboot bars, the text
for app upgrades and the horizontal lines are pretty specific to us.

The current thinking is that just having a canvas with scaling and clipping
should get us quite far.